Prevented crashing on DB down

+++ events/guild/guildDelete.js
@@ -1,6 +1,8 @@
 const serverModel = require("../../models/serverSchema");
 
 module.exports = async (client, guild) => {
+    if(!client.dbCmds) return;
+    
     try {
         await serverModel.findOneAndDelete({ serverId: guild.id });
     } catch(err) {
